rename bridge plugin package
