+++ .dockerignore
@@ -11,5 +11,5 @@ demo-vaults
 data
 tmp
 **/dist
-packages/bridge-plugin/main.js
+packages/bridge/main.js
 apps/ignis-server/server/plugins/*/plugin/main.js
