+++ .gitignore
@@ -3,7 +3,7 @@ dist/
 investigation/
 vaults/
 packages/*/dist/
-packages/bridge-plugin/main.js
+packages/bridge/main.js
 apps/ignis-server/server/plugins/*/plugin/main.js
 demo-vaults/
 data/
