+++ apps/ignis-server/server/bridge-plugin.js
@@ -6,7 +6,7 @@ const {
 } = require("./plugin-system/obsidian-plugin");
 
 const BRIDGE_PLUGIN_ID = "ignis-bridge";
-const BRIDGE_PLUGIN_DIR = path.join(__dirname, "..", "..", "..", "packages", "bridge-plugin");
+const BRIDGE_PLUGIN_DIR = path.join(__dirname, "..", "..", "..", "packages", "bridge");
 
 // .ignis metadata helpers
