+++ packages/bridge/manifest.json
