+++ packages/bridge/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@ignis/bridge-plugin",
+  "name": "@ignis/bridge",
   "version": "0.0.0-internal",
   "private": true,
   "scripts": {
