+++ packages/bridge/src/main.js
