+++ packages/bridge/src/plugin-registry.js
