+++ packages/bridge/src/settings/inject.js
