+++ packages/bridge/src/settings/plugin-tabs.js
