+++ packages/bridge/src/settings/server-plugins-tab.js
