+++ packages/bridge/src/status-bar.js
